+++ src/azure-cli-core/azure/cli/core/__init__.py
@@ -2,7 +2,5 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-import pkg_resources
-pkg_resources.declare_namespace(__name__)
 
 __version__ = "2.0.3+dev"