+++ src/azure-cli-core/azure/cli/core/util.py
@@ -85,6 +85,9 @@ def show_version_info_exit(out_file):
           file=out_file)
     print(file=out_file)
     print('Python ({}) {}'.format(platform.system(), sys.version), file=out_file)
+    print(file=out_file)
+    print("Python location '{}'".format(sys.executable), file=out_file)
+    print(file=out_file)
     sys.exit(0)