+++ src/command_modules/azure-cli-network/azure/cli/command_modules/network/_format.py
@@ -98,10 +98,21 @@ def transform_vpn_connection(result):
                     delattr(prop_val, prop)
     return result
 
+
 def transform_vpn_connection_create_output(result):
     from azure.cli.core.commands import DeploymentOutputLongRunningOperation
-    result = DeploymentOutputLongRunningOperation('Starting network vpn-connection create')(result)
-    return result['resource']
+    from msrest.pipeline import ClientRawResponse
+    from msrestazure.azure_operation import AzureOperationPoller
+    if isinstance(result, AzureOperationPoller):
+        # normally returns a LRO poller
+        result = DeploymentOutputLongRunningOperation('Starting network vpn-connection create')(result)
+        return result['resource']
+    elif isinstance(result, ClientRawResponse):
+        # returns a raw response if --no-wait used
+        return
+    else:
+        # returns a plain response (not a poller) if --validate used
+        return result
 
 
 def transform_vnet_create_output(result):